requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
dnspython==2.4.2
exceptiongroup==1.2.0
frozenlist==1.4.0
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
motor==3.3.1
multidict==6.0.4
psutil==5.9.5
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
textdistance==4.6.0
urllib3==2.0.5
yarl==1.9.2
